Powershell changed, database is being rebuilt on powershell script being run.

+++ DatabaseProject/Data/Sql/Tables/Movie.Showing.sql
@@ -4,7 +4,7 @@ BEGIN
 	(
 		ShowingID INT NOT NULL IDENTITY(1,1) PRIMARY KEY,
 		ShowTime DATETIMEOFFSET NOT NULL,
-		TicketPrice FLOAT(4,2) NOT NULL, -- 4 digits, 2 after the decimal point -> XX.YY
+		TicketPrice FLOAT NOT NULL, -- 4 digits, 2 after the decimal point -> XX.YY
 		MovieID INT NOT NULL
 			REFERENCES Movie.Movie(MovieID)
 	);
